+++ src/org/schema/creativeWork/DataFeed.php
@@ -9,7 +9,7 @@ use org\schema\Thing;
  * A single feed providing structured information about one or more entities or topics.
  * @see https://schema.org/DataFeed
  */
-class DataFeed extends DataSet
+class DataFeed extends Dataset
 {
     /**
      * An item within a data feed. Data feeds may have many elements.